org.glassfish.jersey.Uri refactored to org.glassfish.jersey.uri.Uri
