+++ core-common/src/main/java/org/glassfish/jersey/uri/Uri.java
@@ -15,7 +15,7 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
 
-package org.glassfish.jersey;
+package org.glassfish.jersey.uri;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.Retention;
