+++ examples/rx-client-webapp/src/main/java/org/glassfish/jersey/examples/rx/agent/ListenableFutureAgentResource.java
@@ -30,7 +30,7 @@ import org.glassfish.jersey.examples.rx.domain.Destination;
 import org.glassfish.jersey.examples.rx.domain.Forecast;
 import org.glassfish.jersey.examples.rx.domain.Recommendation;
 import org.glassfish.jersey.server.ManagedAsync;
-import org.glassfish.jersey.Uri;
+import org.glassfish.jersey.uri.Uri;
 
 import com.google.common.collect.Lists;
 import com.google.common.util.concurrent.AsyncFunction;
