+++ tests/e2e/src/test/java/org/glassfish/jersey/tests/api/ResponseE2ETest.java
@@ -40,7 +40,7 @@ import javax.ws.rs.core.Response;
 import javax.ws.rs.ext.RuntimeDelegate;
 
 import org.glassfish.jersey.server.ResourceConfig;
-import org.glassfish.jersey.Uri;
+import org.glassfish.jersey.uri.Uri;
 import org.glassfish.jersey.test.JerseyTest;
 
 import org.junit.Test;
